requirements.txt
# FastAPI Backend for OneStopRadio DJ Platform
# Modern async Python API with JWT authentication, station management, and file uploads

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration & Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8

# HTTP & CORS
httpx==0.25.2

# File handling & validation
Pillow==10.1.0
python-magic==0.4.27

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1